+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project changes will be shown here.
 - Issue 186 - Time limit request bug with IMT team
 - Issue 184 - Request 038 - Daily Open SR Report for customers
 - Issue 193 - Monitor SR feature to receive updates automatically
+- Issue 201 - Create items not yet received process
 
 ### Changed
 - Issue 176 - Change background colours on service request page